terraform/modules/cloudrunfunction/src/requirements.txt
functions-framework==3.1.0
google-auth==2.6.0
requests==2.26.0
google-cloud-pubsub
kfp==2.7.0
